src/main/java/com/systemdesign/Player.java
package com.systemdesign;

public class Player {

    private String name;
    private Symbol symbol;
    private boolean winner;

    public Player(String name, Symbol symbol) {
        this.name = name;
        this.symbol = symbol;
    }

    public String getName() {
        return name;
    }

    public Symbol getSymbol() {
        return symbol;
    }

    public boolean isWinner() {
        return winner;
    }

    public void setWinner(boolean winner) {
        this.winner = winner;
    }
}
